Added db changes

+++ DataModels/Entities/AirCraftEquipment.cs
@@ -40,13 +40,13 @@ namespace DataModels.Entities
         
         public Nullable<System.DateTime> CreatedOn { get; set; }
         
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         
-        public Nullable<int> DeletedBy { get; set; }
+        public Nullable<long> DeletedBy { get; set; }
         
         public Nullable<System.DateTime> DeletedOn { get; set; }
         
-        public Nullable<int> UpdatedBy { get; set; }
+        public Nullable<long> UpdatedBy { get; set; }
         
         public Nullable<System.DateTime> UpdatedOn { get; set; }
     }
