+++ DataModels/VM/AircraftEquipment/AircraftEquipementTimeVM.cs
@@ -18,11 +18,11 @@ namespace DataModels.VM.AircraftEquipment
         public long AircraftScheduleId { get; set; }
         public decimal TotalHours { get; set; }
         public bool IsDeleted { get; set; }
-        public Nullable<int> DeletedBy { get; set; }
+        public Nullable<long> DeletedBy { get; set; }
         public Nullable<System.DateTime> DeletedOn { get; set; }
         public Nullable<System.DateTime> CreatedOn { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> UpdatedOn { get; set; }
-        public Nullable<int> UpdatedBy { get; set; }
+        public Nullable<long> UpdatedBy { get; set; }
     }
 }
