+++ DataModels/VM/User/UserVM.cs
@@ -63,9 +63,9 @@ namespace DataModels.VM.User
         public bool IsActive { get; set; }
         public bool IsDeleted { get; set; }
         public Nullable<System.DateTime> CreatedOn { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> UpdatedOn { get; set; }
-        public Nullable<int> UpdatedBy { get; set; }
+        public Nullable<long> UpdatedBy { get; set; }
 
         public string ActivationLink { get; set; }
